% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map a specific state's accidents in a specific year selected.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A non-negative interger specifying the desired state}

\item{year}{A non-negative interger specifying the desired year.}
}
\value{
A map of a state marked accidents as circle
}
\description{
Map a specific state's accidents in a specific year selected.
}
\examples{
library(maps)
library(graphics)
\dontrun{fars_map_state(1, 2013)}


}
\seealso{
\code{\link{fars_read}} \code{\link{make_filename}} \code{\link{make_filename}} \code{\link{make_filename}}
}
