% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a CSV file and create it a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A charater string of the name of file to be imported or a relative path to a file.
If file does not exist in the working directory or specific directory, function will be stopped with an error message.}
}
\value{
Only prints the first 10 rows and the columns that fit on screen. The rest of columns which were not shown in tbl are listed below tbl as text.
}
\description{
This is a simple function that reads a csv(common separated values) file and create a data frame from it,
with cases corresponding to rows(lines) and variables to columns(fields) in the file. tbl objects only print
a few rows and all the columns that fit on one screen, describing the rest of it as text.
}
\examples{
fars_read("extdata/accident_2013.csv")

}
