% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a specific year file and Select MONTH, year columns of it.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A non-negative interger specifying the desired year. This year's data will be loaded in a tbl by \link{fars_read}}
}
\value{
A specific columns of loaded data. The MONTH, YEAR columns will be selected
}
\description{
Read a desired year's pre-named file that was made by \link{make_filename} function
             and Select only MONTH, year two columns matching year value in the tbl with the your input year value.
             If you input unexist year, it will be error due to there is no file to import
}
\examples{
\dontrun{fars_read_years(2013)}
\dontrun{fars_read_years("2013")}

}
