% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Sum count by month of the year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years values (A non-negative interger specifying the desired year)}
}
\value{
A tbl including MONTH and it's count value. Total 12 rows and month and count columns will be shown.
}
\description{
\code{fars_summarize_year} reads a desired years pre-named file by using \code{\link{fars_read_years}} function and group by year, MONTH.
             And then summarize counting by month of the year by \code{\link[dplyr]{summarize}}
}
\examples{
library(tidyr)
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years("2013")}
\dontrun{fars_summarize_years(c("2013","2014")}

}
\seealso{
\code{\link{fars_read_years}}
}
