% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Write file name attaching desired year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A non-negative integer specifying the desired year.}
}
\value{
A format string including a year, such as accident_2017.csv.bz2
}
\description{
This is a simple function that makes file name as pre-defined format attaching a specific year
              parameter of an argument
}
\examples{
make_filename(2017)
make_filename("2017")

}
